+++ packages/rma-server/src/warranty-claim/entity/warranty-claim/warranty-claim-dto.ts
@@ -186,4 +186,10 @@ export class WarrantyClaimDto {
 
   @IsOptional()
   parent: string;
+
+  @IsOptional()
+  service_vouchers: string[];
+
+  @IsOptional()
+  service_items: string[];
 }