+++ packages/rma-server/src/warranty-claim/entity/warranty-claim/warranty-claim.entity.ts
@@ -157,6 +157,12 @@ export class WarrantyClaim extends BaseEntity {
 
   @Column()
   parent: string;
+
+  @Column()
+  service_vouchers: string[];
+
+  @Column()
+  service_items: string[];
 }
 
 export class WarrantyBulkProducts {