+++ packages/rma-warranty/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://gitlab.com/castlecraft/excel-rma/compare/rma-warranty@1.4.0...rma-warranty@1.4.1) (2021-02-21)
+
+**Note:** Version bump only for package rma-warranty
+
+
+
+
+
 # [1.4.0](https://gitlab.com/castlecraft/excel-rma/compare/rma-warranty@1.3.0...rma-warranty@1.4.0) (2021-02-18)